third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-paint/timeline-paint-with-style-recalc-invalidations-expected.txt

Tests the Timeline API instrumentation of paint events with style recalc invalidations.

PASS


Running: testLocalFrame
first paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-style-recalc-invalidations.html:9}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]

Running: testSubframe
second paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: http://127.0.0.1:8000/devtools/tracing/timeline-paint/timeline-paint-with-style-recalc-invalidations.html:16}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
]
